output mode as integer not as char

+++ src/surf_query.cpp
@@ -177,7 +177,7 @@ int main(int argc,char* const argv[])
                 std::cout << surf_resp->qry_id << ";" 
                           << surf_resp->collection << ";"
                           << surf_resp->index << ";"
-                          << qry_mode << ";"
+                          << (int)qry_mode << ";"
                           << surf_resp->k << ";"
                           << surf_resp->qry_len << ";"
                           << surf_resp->result_size << ";"
